backend/requirements.txt
Flask
Flask-Cors
celery
redis
python-dotenv
Flask-Limiter
